Cursed hack to avoid horizontal layout shift
The problem is that when the stat text gets long, it decides the width of the parent container. This is problematic as the transition animation will gradually write the stat text causing weird jitters with the width of the container as the text writes and wraps.

By adding `w-0` and `min-w-full` to the stat component, it will limit its width to that of the container. However, the width of the container is subject to change because the bio's width also changes. Adding the same two classes to the bio as well makes both children take the width of the longest child which is the name.

However, I want the bio to take the width it needs and not be scrunched up when it doesn't need to be.

To fix this, I added an element with a height of 0 and static pixel widths matching that of the bio at each of its breakpoints. This way, the bio and stats will then try to be the width of this element.